% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_composite_score}
\alias{calculate_composite_score}
\title{Calculate composite scores, ranks, and quantiles from domain scores. This
function will calculate over all numeric variables in a dataframe.}
\usage{
calculate_composite_score(data, index_name, num_quantiles = 10)
}
\arguments{
\item{data}{Data frame containing domain scores}

\item{index_name}{A string identifier to prefix to column names. Use
snake_case.}

\item{quantiles}{The Number of quantiles}
}
\description{
Calculate composite scores, ranks, and quantiles from domain scores. This
function will calculate over all numeric variables in a dataframe.
}
