% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_domain_scores}
\alias{calculate_domain_scores}
\title{Calculate domain scores, ranks, and quantiles from normalised indicators. The
indicators can be weighted or unweighted. This function will calculate over
all numeric variables in a dataframe.}
\usage{
calculate_domain_scores(
  data,
  domain_name,
  num_quantiles = 10,
  ignore_nas = FALSE
)
}
\arguments{
\item{data}{Data frame containing weighted indicators}

\item{domain_name}{A string identifier to prefix to column names. Use
snake_case.}

\item{ignore_nas}{Whether should ignore missing values when summing across
the row. The default is FALSE so flags these values.}

\item{quantiles}{The Number of quantiles}
}
\description{
Calculate domain scores, ranks, and quantiles from normalised indicators. The
indicators can be weighted or unweighted. This function will calculate over
all numeric variables in a dataframe.
}
