% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dftest}
\alias{dftest}
\title{DF}
\format{
A data frame with .. rows and . variables:
\describe{
\item{X1}{...}
\item{X1.1}{..}
...
}
}
\source{
\url{https://www..../}
}
\usage{
dftest
}
\description{
A dataset ..
}
\keyword{datasets}
