% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_file}
\alias{download_file}
\title{Download a file temporarily to disk}
\usage{
download_file(url, file_extension)
}
\arguments{
\item{url}{A URL for the request}

\item{file_extension}{A character vector detailing the file extension (e.g.,
".xlsx")}
}
\description{
Download a file temporarily to disk
}
