% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_codes}
\alias{filter_codes}
\title{Filter geographical codes by a regex pattern}
\usage{
filter_codes(.data, codes, pattern)
}
\arguments{
\item{.data}{A data frame}

\item{codes}{The column containing geographical codes}

\item{pattern}{Pattern to look for. Use a regular expression}
}
\description{
Filter geographical codes by a regex pattern
}
