% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_adult_overweight_obese}
\alias{hl_adult_overweight_obese}
\title{Percentage of Overweight and Obese Adults (aged 16+) in Wales}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{Percentage_overweight_obese}{The percentage of overweight and obese adults in each local authority of Wales}
\item{Year}{Time frame}
.  ...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/National-Survey-for-Wales/Population-Health/Adult-Lifestyles}
}
\usage{
hl_adult_overweight_obese
}
\description{
Dataset containing information about the percentage of overweight and obese adults in Wales aged 16+
}
\keyword{datasets}
