% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_alcohol_misuse}
\alias{hl_alcohol_misuse}
\title{Alcohol Death Rates 2020-2022}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Local Authority area code}
\item{alcohol_death_rate_per_100k}{Age standardised alcohol-specific death rate per 100,000}
\item{Year}{Time frame}
.  ...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/healthandsocialcare/causesofdeath/datasets/alcoholspecificdeathsintheukmaindataset}
}
\usage{
hl_alcohol_misuse
}
\description{
Dataset containing information about death rate from alcohol misuse in Wales
}
\keyword{datasets}
