% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_cancer_screening_bowel}
\alias{hl_cancer_screening_bowel}
\title{Bowel Cancer Screening Attendance Wales 2021-2022}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{Screening uptake percentage}{Percentage of adults out of those invited to attend bowel cancer screening (aged 58-74) who attended their screening. Participants were deemed to have responded to their invitation if the bowel screening programme received a used test kit within six months following their invitation.}
\item{Year}{Financial Year}
.  ...
}
}
\source{
\url{https://phw.nhs.wales/services-and-teams/screening/bowel-screening/programme-reports/uptake-coverage-by-local-authority-and-health-boards/}
}
\usage{
hl_cancer_screening_bowel
}
\description{
Dataset containing information about the percentage of adults age 58-74 who attended bowel cancer screenings in Wales
}
\keyword{datasets}
