% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_cancer_screening_breast}
\alias{hl_cancer_screening_breast}
\title{Breast Cancer Screening Attendance Wales 2021-2022}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{Screening uptake percentage}{Percentage of adults out of those invited to attend breast cancer screening (aged 50-70) who attended their screening. Women were counted as having responded to their invitation if they attended for screening within six months of original invitation.}
\item{Year}{Time frame for last recorded Breast Cancer screening, May 2021}
.  ...
}
}
\source{
\url{https://phw.nhs.wales/services-and-teams/screening/breast-screening/programme-reports/uptake-coverage-by-local-authority-and-health-boards/}
}
\usage{
hl_cancer_screening_breast
}
\description{
Dataset containing information about the percentage of women age 50-70 who attended breast cancer screenings
}
\keyword{datasets}
