% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_cancer_screening_cervical}
\alias{hl_cancer_screening_cervical}
\title{Cervical Cancer Screening Attendance Wales 2021-2022}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{Screening uptake percentage}{Percentage of adults out of those invited to attend cervical cancer screening (aged 25-64) who attended their screening. Women were counted as having responded if they are aged 25-49 years who have received an adequate test in the last 3.5 years and if they are aged 50-64 years who received an adequate test in the last 5.5 years.}
\item{Year}{Financial Year}
.  ...
}
}
\source{
\url{https://phw.nhs.wales/services-and-teams/cervical-screening-wales/information-resources/programme-reports/uptake-coverage-by-local-authority-and-health-boards/}
}
\usage{
hl_cancer_screening_cervical
}
\description{
Dataset containing information about the percentage of women age 25-64 who attended cervical cancer screenings
}
\keyword{datasets}
