% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_drug_misuse}
\alias{hl_drug_misuse}
\title{Drug Poisoning Death Rates 2022}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code of the local authority}
\item{Drug poisoning death rate}{Drug poisoning related deaths per 1,000 people}
\item{Year}{Year of the data}
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/deaths/datasets/drugmisusedeathsbylocalauthority}

\url{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/estimatesofthepopulationforenglandandwales}
}
\usage{
hl_drug_misuse
}
\description{
Dataset containing information about drug poisoning death rates in Wales for the year 2022.
}
\keyword{datasets}
