% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_education_employment_apprenticeship}
\alias{hl_education_employment_apprenticeship}
\title{Young people in education, employment and apprenticeships Wales 2009/2010}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{participation_rate_eea}{Participation rate in post-16 (after Key Stage 4)
Education and Training in Wales for under 20 year olds. Participation rate
is calculated using census data population count by local authority,
with participation measured against the Welsh national average of 100.
Participation rates above 100 reflect high participation rates, below 100
low participation rates}
\item{Year}{Academic year}
.  ...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Education-and-Skills/Post-16-Education-and-Training/Further-Education-and-Work-Based-Learning/Standardised-Participation-Rates}
}
\usage{
hl_education_employment_apprenticeship
}
\description{
Dataset containing information about the standard participation rate in sustained education,
employment or an apprenticeship for 16-24 year olds after key stage 4
}
\keyword{datasets}
