% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_healthy_eating}
\alias{hl_healthy_eating}
\title{Healthy eating Wales 2021-2023}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{Percent adults who ate five fruit/veg yesterday}{Percentage of people aged 16 and over in each local authority claiming to have eaten at least 5 portions of fruit and vegetables the previous day, age standardised}
\item{Year}{Time frame}
.  ...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/National-Survey-for-Wales/Population-Health/Adult-Lifestyles}
}
\usage{
hl_healthy_eating
}
\description{
Dataset containing information about healthy eating in Wales
}
\keyword{datasets}
