% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_low_birth_weight}
\alias{hl_low_birth_weight}
\title{Low Birth Weight Percentage by year Wales}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{percentage_low_birth_weight}{The percentage of singleton live births under 2500g in each local authority of Wales}
\item{Year}{Time frame}
.  ...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/National-Survey-for-Wales/Population-Health/Adult-Lifestyles}
}
\usage{
hl_low_birth_weight
}
\description{
Dataset containing information about the percentage of live births under 2500g in Wales
}
\keyword{datasets}
