% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_mmr_vaccination_coverage}
\alias{hl_mmr_vaccination_coverage}
\title{MMR vaccination coverage Wales 2021-2023}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{MMR percentage coverage by 2nd birthday}{Percentage of children immunised against Measles, Mumps and Rubella (MMR) reaching their second birthday in the year}
\item{Year}{Time frame}
.  ...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Health-and-Social-Care/NHS-Primary-and-Community-Activity/Immunisation}
}
\usage{
hl_mmr_vaccination_coverage
}
\description{
Dataset containing information about the percentage of children in Wales immunized against Measles, Mumps and Rubella (MMR) by their second birthday
}
\keyword{datasets}
