% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_primary_pupil_absences}
\alias{hl_primary_pupil_absences}
\title{Primary Pupil Absences Wales 2022/2023}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{primary_school_absence_percentage}{Percentage of half day sessions missed
for primary school students, aged 5-10, for the full academic year, in Wales}
\item{year}{Academic year}
.  ...
}
}
\source{
\url{https://www.gov.wales/absenteeism-primary-schools-september-2022-august-2023}
}
\usage{
hl_primary_pupil_absences
}
\description{
Dataset containing information about the percentage of absences for primary
school students in Wales for the academic year 2022/2023
}
\keyword{datasets}
