% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_sedentary_behaviour}
\alias{hl_sedentary_behaviour}
\title{Smoking Wales 2021-2023}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{Percent smokers}{Percentage of people aged 16 and over in each local authority who are self-reported smokers, age standardised}
\item{Year}{Time frame}
.  ...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/National-Survey-for-Wales/Population-Health/Adult-Lifestyles}
"hl_smoking"
Sedentary behaviour Wales 2021-2023

Dataset containing information about the percentage of sedentary adults in Wales
\item{Percent adults active less than 30 minutes last week}{Percentage of people aged 16 and over in each local authority claiming to have exercised for less than 30 minutes the previous week, age standardised}
\item{Year}{Time frame}
.  ...
}

\url{https://statswales.gov.wales/Catalogue/National-Survey-for-Wales/Population-Health/Adult-Lifestyles}
}
\usage{
hl_sedentary_behaviour
}
\description{
Dataset containing information about the percentage of smokers in Wales
}
\keyword{datasets}
