% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hl_smoking}
\alias{hl_smoking}
\title{Smoking Wales 2021-2023}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{Percent smokers}{Percentage of people aged 16 and over in each local authority who are self-reported smokers, age standardised}
\item{Year}{Time frame}
.  ...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/National-Survey-for-Wales/Population-Health/Adult-Lifestyles}
}
\usage{
hl_smoking
}
\description{
Dataset containing information about the percentage of smokers in Wales
}
\keyword{datasets}
