% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hp_greenspace_access}
\alias{hp_greenspace_access}
\title{Greenspace Provision in Wales 2024}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Area code}
\item{ltla21_name}{Area name}
\item{provision_per_person}{The amount of public greenspace available per person, SQM}
.  ...
}
}
\source{
\url{https://fieldsintrust.org/insights/green-space-index}
}
\usage{
hp_greenspace_access
}
\description{
Dataset containing information about greenspace provision in Wales
}
\keyword{datasets}
