% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hp_household_overcrowding}
\alias{hp_household_overcrowding}
\title{Percentage of Households Experiencing Overcrowding in Wales (2021 Census)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla21_code}{Local Authority Code}
\item{percentage_households_overcrowding}{Percentage of households assessed
as overcrowded at the time of the 2021 Census, per Local Authroity}
\item{year}{Date}

...
}
}
\source{
\url{https://www.ons.gov.uk/file?uri=/peoplepopulationandcommunity/housing/datasets/overcrowdingandunderoccupancybyhouseholdcharacteristicsenglandandwalescensus2021/census2021/hou04dataset.xlsx}
}
\usage{
hp_household_overcrowding
}
\description{
A dataset containing statistics on the percentage of households in each LA that are experiencing overcrowding, 2021.
}
\details{
Overcrowding is defined as households with an occupancy rating of -1 or less,
which implies that a household has at least one fewer bedroom than required.
}
\keyword{datasets}
