% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hp_rough_sleepers}
\alias{hp_rough_sleepers}
\title{Wales rough sleepers - (2023)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla22_name}{Local Authority Code}
\item{rough_sleepers_per_1000}{People assessed as rough sleepers per 1,000
individuals in the Local Authority}
\item{year}{Date}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Housing/Homelessness/homelessness-accommodation-provision-and-rough-sleeping/roughsleepers-by-localauthority}
}
\usage{
hp_rough_sleepers
}
\description{
A dataset containing statistics on rates of rough sleepers by 1,000, by
Local Authority
}
\keyword{datasets}
