% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inverse_rank}
\alias{inverse_rank}
\title{Inverse ranking with NAs first (i.e. 1 = best)}
\usage{
inverse_rank(x)
}
\arguments{
\item{x}{Data to rank}
}
\description{
Inverse ranking with NAs first (i.e. 1 = best)
}
