% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{invert_this}
\alias{invert_this}
\title{Invert a vector (e.g., deciles, ranks, percentiles, etc.)
For example, with deciles, a score of 10 -> 1 and a score of 1 -> 10}
\usage{
invert_this(x)
}
\arguments{
\item{x}{Vector of data to invert}
}
\description{
Invert a vector (e.g., deciles, ranks, percentiles, etc.)
For example, with deciles, a score of 10 -> 1 and a score of 1 -> 10
}
