% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_alcohol_misuse}
\alias{lives_alcohol_misuse}
\title{Rate of Patient Admissions for Alcohol Related Conditions per 100k (2023/24)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{alcohol_admissions_rate_per_100k}{Rate of patient admissions per 100k
population, based on population numbers per Local Authority}
\item{year}{Time period}

...
}
}
\source{
\url{https://phw.nhs.wales/publications/publications1/data-mining-wales-the-annual-profile-for-substance-misuse-2023-2024/}
}
\usage{
lives_alcohol_misuse
}
\description{
A dataset containing statistics on the rate of patient admissions for
all alcohol-related admissions per 100k in each Welsh Council, 2023/24.
}
\keyword{datasets}
