% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_cancer_screening}
\alias{lives_cancer_screening}
\title{Percentage of Cancer Screening Uptake (2021-22)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla23_code}{Local Authority Code}
\item{total_cancer_screening_percentage}{Percentage of cancer screening
uptake; breast, bowel, and cervical cancer.}
\item{year}{Time period}

...
}
}
\source{
\url{https://phw.nhs.wales/services-and-teams/screening/bowel-screening/programme-reports/uptake-coverage-by-local-authority-and-health-boards/}

\url{https://phw.nhs.wales/services-and-teams/cervical-screening-wales/information-resources/programme-reports/uptake-coverage-by-local-authority-and-health-boards/}

\url{https://phw.nhs.wales/services-and-teams/screening/breast-screening/programme-reports/uptake-coverage-by-local-authority-and-health-boards/}
}
\usage{
lives_cancer_screening
}
\description{
A dataset containing statistics on the percentage of cancer screening uptake
for bowel, breast, and cervical cancer in each Welsh Council, 2021-22.
}
\keyword{datasets}
