% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_child_vaccination_coverage}
\alias{lives_child_vaccination_coverage}
\title{Percentage of Childhood Vaccine Coverage (2023-2024)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{child_vaccination_coverage_percentage}{Percentage of vaccine coverage.
Includes the following immunisations: Diphtheria, Tetanus, Pertussis, Polio,
Haemophilus influenzae (Hib), Measles, Mumps, Rubella, Meningitis C and
Pneumococcal infection (PCV).}
\item{year}{School Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Health-and-Social-Care/NHS-Primary-and-Community-Activity/flying-start/childrenlivinginflyingstartnonflyingstartareaswhoarefullyimmunisedbytheir4thbirthday-by-localauthority}
}
\usage{
lives_child_vaccination_coverage
}
\description{
A dataset containing statistics on childhood vaccine coverage (aged 4 and
under) in each Welsh Council, 2023/24.
}
\keyword{datasets}
