% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_drug_misuse}
\alias{lives_drug_misuse}
\title{Drug misuse per 1k (2024)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{drug_misuse_per_1k}{Average rate of drug offences per 1,000 people}
\item{year}{Year}
...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/crimeandjustice/datasets/policeforceareadatatables}
}
\usage{
lives_drug_misuse
}
\description{
A dataset containing statistics on the number of drug-related crimes recorded
by the police, per 1,000 people, per Welsh Council (2024). Note, that these
data have been aggregated up from the four police areas: Dyfed-Powys, Gwent
and North and South Wales, and so repeated obersations will be seen across
local authorities.
}
\keyword{datasets}
