% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_healthy_eating}
\alias{lives_healthy_eating}
\title{Percentage of People that Eat Healthy (2021-22 and 2022-23)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla23_code}{Local Authority Code}
\item{healthy_eating_percentage}{Percentage of people that consume
5 portions or more of fruit and vegetable per day}
\item{year}{Time periods combined}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/National-Survey-for-Wales/Population-Health/Adult-Lifestyles/adultlifestyles-by-healthboard-from-202021}
}
\usage{
lives_healthy_eating
}
\description{
A dataset containing statistics on fruit and vegetable consumption in
each Welsh Council, 2021-22 and 2022-23.
}
\details{
To note: England's Health Index captures percentage of healthy eating per
year, but Wales only has data combining over a 2 year period.
}
\keyword{datasets}
