% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_low_birth_weight}
\alias{lives_low_birth_weight}
\title{Percentage of Babies Born Not at a Healthy Birth Weight (2023)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{low_birth_weight_percentage}{Percentage of babies born not at a healthy
birth weight (<2,500g)}
\item{year}{Year}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Health-and-Social-Care/NHS-Primary-and-Community-Activity/Community-Child-Health/livebirthstowelshresidents-by-localhealthboard-birthweight}
}
\usage{
lives_low_birth_weight
}
\description{
A dataset containing percentage babies born not at a healthy birth weight
(<2,500g).
}
\keyword{datasets}
