% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_obesity_adults}
\alias{lives_obesity_adults}
\title{Average percentage of Adults Obese (2024)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{obesity_adults_percentage}{Average percentage of adults
(aged 16+) clinically classed as overweight and obese}
\item{year}{Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Health-and-Social-Care/NHS-Primary-and-Community-Activity/GMS-Contract/diseaseregisters-by-localhealthboard}
}
\usage{
lives_obesity_adults
}
\description{
A dataset containing statistics on the average percentage of adults (aged
16+) clinically classed as obese in each Council, 2024.
}
\details{
To note: England's Health Index captures data on overweight and obese adults
aged 16+. However, Wales only captures adults classed as obese. This has been
included as the closest proxy.

To note: StatsWales Disease Prevalence data only captures disease prevalence
in a 'snapshot', or a single point of time. This means the year data
represents when the data was extracted, not the period covered.
}
\keyword{datasets}
