% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_overweight_obesity_children}
\alias{lives_overweight_obesity_children}
\title{Average percentage of Children Overweight/Obese (2022/23)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{overweight_obesity_child_percentage}{Average percentage of children
clinically classed as overweight and obese aged 4/5}
\item{year}{School Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Health-and-Social-Care/NHS-Primary-and-Community-Activity/flying-start/prevalenceofchildrenin-healthy-weight-categoriesforchildrenaged4to5yearsresidentwithinflyingstartandnonflyingstartareas-by-localauthority}
}
\usage{
lives_overweight_obesity_children
}
\description{
A dataset containing statistics on the average percentage of children
clinically classed as overweight and obese in each Council, 2022/23.
}
\details{
To note: England's Health Index captures the percentage of children aged 4-5
and 10-11 respectively who are classed as being overweight or obese. However,
the only available Welsh data captured overweight and obesity data for 4-5
year olds, so this has been included as the closest proxy here.
}
\keyword{datasets}
