% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_physical_activity}
\alias{lives_physical_activity}
\title{Percentage of People Meeting Recommended Activity Levels (2021-22 and 2022-23)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{activity_levels_met_percentage}{Percentage of people that meet the
recommended activity levels of at least 150 minutes a week}
\item{year}{Time periods over a 2 year period}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/National-Survey-for-Wales/Population-Health/Adult-Lifestyles/adultlifestyles-by-healthboard-from-202021}
}
\usage{
lives_physical_activity
}
\description{
A dataset containing statistics on the percentage of people that meet the
recommended activity levels of at least 150 minutes a week in each Welsh
Council, 2021-22 and 2022-23.
}
\keyword{datasets}
