% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_pupil_absence}
\alias{lives_pupil_absence}
\title{Percentage of Pupil Absences (2023/24)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{total_pupil_absence_percentage}{Percentage of unauthorised school
absences across primary and secondary schools}
\item{year}{School Year}

...
}
}
\source{
\url{https://www.gov.wales/absenteeism-primary-schools-september-2023-august-2024}

\url{https://www.gov.wales/absenteeism-secondary-schools-september-2023-august-2024}
}
\usage{
lives_pupil_absence
}
\description{
A dataset containing statistics on the percentage of unauthorised school
absences.
}
\details{
To note: England's Health Index records the percentage of pupils (at
state-funded primary and secondary, and special schools) who are persistent
absentees, that is, have overall absences equating to 10\% or more of their
possible sessions. However, Wales only records percentage of half-day
sessions missed for unauthorised absence for primary and secondary schools.
This has been used as best proxy.
}
\keyword{datasets}
