% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_sedentary_behaviour}
\alias{lives_sedentary_behaviour}
\title{Percentage of People Exhibiting Sedentary Behaviour (Very Low Activity Levels)
(2021-22 and 2022-23)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{sedentary_behaviour_percentage}{Percentage of people with very low
activity levels, defined as less than 30 minutes activity a week}
\item{year}{Time periods over a 2 year period}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/National-Survey-for-Wales/Population-Health/Adult-Lifestyles/adultlifestyles-by-healthboard-from-202021}
}
\usage{
lives_sedentary_behaviour
}
\description{
A dataset containing statistics on very low activity levels (less than 30
minutes a week) in each Council, 2021-22 and 2022-23.
}
\keyword{datasets}
