% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_sexual_health}
\alias{lives_sexual_health}
\title{Sexual health Rates per 100k (2023)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{sexual_health_testing_diagnosis_rate_per_100k}{Rate of chlamydia,
gonorrhoea, syphilis, and testing rates per 100,000}
\item{year}{Year}

...
}
}
\source{
\url{https://phw.nhs.wales/publications/publications1/sexual-health-trends-in-wales-annual-report-2024/}
}
\usage{
lives_sexual_health
}
\description{
A dataset containing statistics on rates of chlamydia, gonorrhoea, syphilis,
and testing rates per 100k in each Welsh Council, 2023.
}
\keyword{datasets}
