% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_teenage_pregnancy}
\alias{lives_teenage_pregnancy}
\title{Rates of teenage pregnancies per 1,000 (2021)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla25_code}{Local Authority Code}
\item{teenage_pregnancies_per_1k}{Rate of teenage pregnancies per 1k where
teenagers are aged between 13 and 18}
\item{year}{Year}

...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/conceptionandfertilityrates/datasets/conceptionstatisticsenglandandwalesreferencetables}
}
\usage{
lives_teenage_pregnancy
}
\description{
A dataset containing statistics on the rate of teenage pregnancies (under 18)
per 1k.
}
\keyword{datasets}
