% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_indicators}
\alias{load_indicators}
\title{Load indicators saved as .rds files within a folder into a tibble, joined by
a key}
\usage{
load_indicators(path_pattern, key)
}
\arguments{
\item{key}{The key to join the variables}

\item{path}{Relative path to the folder containing the .rds files to be read}
}
\description{
Load indicators saved as .rds files within a folder into a tibble, joined by
a key
}
