% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalise}
\alias{normalise}
\title{Normalise a vector where mean = 0 & SD = 1.}
\usage{
normalise(x, ignore_nas = FALSE)
}
\arguments{
\item{x}{Vector of data to normalise}

\item{ignore_nas}{Whether should ignore missing values when normalising. The
default is FALSE so flags these values.}
}
\description{
Normalise a vector where mean = 0 & SD = 1.
}
