% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalise_indicators}
\alias{normalise_indicators}
\title{Normalise indicators to Mean = 0, SD = 1. This function will calculate over
all numeric variables in a dataframe.}
\usage{
normalise_indicators(data, ignore_nas = FALSE)
}
\arguments{
\item{data}{Data frame containing indicators to normalise.}

\item{ignore_nas}{Whether should ignore missing values when normalising. The
default is FALSE so flags these values and is an input in the
normalise function.}
}
\description{
Normalise indicators to Mean = 0, SD = 1. This function will calculate over
all numeric variables in a dataframe.
}
