% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_all_mortality}
\alias{people_all_mortality}
\title{Mortality all causes rate per 100k (2023)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{all_deaths_per_100k}{age-sex standardised rate for all causes of
mortality per 100k}
\item{year}{Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}
...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/deaths/datasets/deathsregisteredinenglandandwalesseriesdrreferencetables}
}
\usage{
people_all_mortality
}
\description{
A dataset containing statistics from an age-sex standardised rate for all
causes of mortality per 100k, by Council (2023).
}
\keyword{datasets}
