% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_anxiety}
\alias{people_anxiety}
\title{Average Measurement of Anxiety Out of 10 (2022-23)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{anxiety_score_out_of_10}{Average personal ratings on feelings of
anxiety out of 10 - 10 is most anxious, 1 is least anxious}
\item{year}{Time period}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}
...
}
}
\source{
\url{https://www.ons.gov.uk/datasets/wellbeing-local-authority/editions/time-series/versions/4}
}
\usage{
people_anxiety
}
\description{
A dataset containing statistics of personal ratings on feelings of
anxiety out of 10, by Council (2022-23).
}
\keyword{datasets}
