% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_avoidable_mortality}
\alias{people_avoidable_mortality}
\title{Rate of Avoidable Deaths per 100k (2020-2022)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{avoidable_deaths_per_100k}{Age-standardised avoidable mortality rates per
100k}
\item{year}{3-year aggregate period}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/healthandsocialcare/causesofdeath/datasets/avoidablemortalitybylocalauthorityinenglandandwales}
}
\usage{
people_avoidable_mortality
}
\description{
A dataset containing statistics on age-standardised death rates per 100k, by
Welsh Council (2020-2022).
}
\keyword{datasets}
