% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_diabetes}
\alias{people_diabetes}
\title{Percentage Rate of Diabetes (Diabetes mellitus (patients aged 17+)) (2024)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla25_code}{Local Authroity Code}
\item{diabetes_percentage}{Percentage of people who
self-reported as having Diabetes mellitus}
\item{year}{Snapshot of a single point in time in 2024 - specifically 1
April 2024}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Health-and-Social-Care/NHS-Primary-and-Community-Activity/GMS-Contract/diseaseregisters-by-localhealthboard}
}
\usage{
people_diabetes
}
\description{
A dataset containing statistics on the percentage of people who self-reported
as having Diabetes mellitus (aged 17+) in Welsh Council Areas.
}
\details{
To note: StatsWales Disease Prevalence data only captures disease prevalence
in a 'snapshot', or a single point of time. This means the year data
represents when the data was extracted, not the period covered.
}
\keyword{datasets}
