% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_life_expectancy}
\alias{people_life_expectancy}
\title{Average Life Expectancy for Men and Women (2021-2023)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{life_expectancy_combined}{Average life expectancy for men and women}
\item{year}{Time period - three year aggregate}

...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/healthandsocialcare/healthandlifeexpectancies/datasets/lifeexpectancyforlocalareasofgreatbritain}
}
\usage{
people_life_expectancy
}
\description{
A dataset containing statistics on average life expectancy for men
and women, by Welsh Council (2021-2023).
}
\keyword{datasets}
