% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_life_satisfaction}
\alias{people_life_satisfaction}
\title{Average Measurement of Life Satisfaction Out of 10 (2022-23)}
\format{
A data frame with 11 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{satisfaction_score_out_of_10}{Average personal ratings on feelings of
life satisfaction out of 10 - 10 is most satisfied, 1 is least satisfied}
\item{year}{Time period}
...
}
}
\source{
\url{https://www.ons.gov.uk/datasets/wellbeing-local-authority/editions/time-series/versions/4}
}
\usage{
people_life_satisfaction
}
\description{
A dataset containing statistics of personal ratings on feelings of
life satisfaction out of 10, by Council (2022-23).
}
\keyword{datasets}
