% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_musculoskeletal_conditions}
\alias{people_musculoskeletal_conditions}
\title{Percentage Rate of Musculoskeletal Conditions (Rheumatoid Arthritis
(patients aged 16+)) (2024)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authroity Code}
\item{musculoskeletal_conditions_percentage}{Percentage of people who
self-reported as having Rheumatoid Arthritis}
\item{year}{Snapshot of a single point in time in 2024 - specifically 1
April 2024}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Health-and-Social-Care/NHS-Primary-and-Community-Activity/GMS-Contract/diseaseregisters-by-localhealthboard}
}
\usage{
people_musculoskeletal_conditions
}
\description{
A dataset containing statistics on the percentage of people who self-reported
as having Rheumatoid Arthritis (aged 16+) in Welsh Council Areas.
}
\details{
To note: StatsWales Disease Prevalence data only captures disease prevalence
in a 'snapshot', or a single point of time. This means the year data
represents when the data was extracted, not the period covered.
}
\keyword{datasets}
