% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{people_suicide}
\alias{people_suicide}
\title{Death Rate per 100k from Suicide (2021-2023)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{suicide_rate_per_100k}{Deaths from suicide (16+). Age standardised
rate per 100,000.}
\item{year}{Time period - 3 year aggregate}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/deaths/datasets/suicidesbylocalauthority}
}
\usage{
people_suicide
}
\description{
A dataset containing statistics on death rate per 100k by suicide, by Welsh
Council.
}
\keyword{datasets}
