% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_air_pollution}
\alias{places_air_pollution}
\title{Air pollution (2019)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{air_pollution_weighted}{Population-weighted annual mean PM2.5}
\item{year}{Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://uk-air.defra.gov.uk/data/pcm-data}
}
\usage{
places_air_pollution
}
\description{
A dataset containing population-weighted annual mean PM2.5 data for 2019, by
Welsh Council Area. Uses the anthropogenic component.
}
\keyword{datasets}
