% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_child_poverty}
\alias{places_child_poverty}
\title{Percentage of Absolute Child Poverty (2022)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla25_code}{Local Authority Code}
\item{child_poverty_percentage}{Percentage of children (aged 16 and under)
living in absolute low income families}
\item{year}{Date}

...
}
}
\source{
\url{https://www.gov.uk/government/statistics/children-in-low-income-families-local-area-statistics-2014-to-2022}
}
\usage{
places_child_poverty
}
\description{
A dataset containing statistics on the percentage of children under 16
experiencing absolute poverty, living in absolute low income families, by
Local Authority, 2022.
}
\details{
Absolute poverty refers to people living in households with income below 60\%
of median income in a base year, usually 2010/11.
}
\keyword{datasets}
