% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_gp_travel_time_msoa}
\alias{places_gp_travel_time_msoa}
\title{Travel time to GPs (2025)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{msoa21_code}{MSOA Code}
\item{gp_mean_travel_time}{Mean travel time (in minutes) to the nearest GP.
999 is a pseudo-value to indicate there are no GPs within 3 hours of the MSOA.}
\item{is_within_3_hours}{Is there a GP within three hours of the centre of the MSOA?}
\item{year}{Year the data was last updated}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}
...
}
}
\source{
\url{https://openstreetmap.org/}
}
\usage{
places_gp_travel_time_msoa
}
\description{
A dataset containing the average (median) travel time (in minutes) to the nearest
GP in each MSOA.
}
\details{
Travel times are calculated using the TravelTime API (https://traveltime.com/apis/distance-matrix)
and are based on travelling by public transport on a weekday morning.
}
\keyword{datasets}
