% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_household_overcrowding}
\alias{places_household_overcrowding}
\title{Percentage Household Overcrowding (2021)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{percentage_households_overcrowding}{Percentage of households
experiencing household overcrowding}
\item{year}{Year}

...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/housing/datasets/overcrowdingandunderoccupancybyhouseholdcharacteristicsenglandandwalescensus2021}
}
\usage{
places_household_overcrowding
}
\description{
A dataset containing statistics on the percentage of households experiencing
household overcrowding, per Council (2021).
Household overcrowding is  defined as households with an occupancy rating of
-1 or lower.
}
\keyword{datasets}
