% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_low_level_crime}
\alias{places_low_level_crime}
\title{Average rates of Low-Level Crime per 1k (2023-2024)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{low_level_crime_rate_per_1k}{Average rate of low-level-crime (bicycle
theft and shoplifting)}
\item{year}{September 2023 - September 2024}
...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/crimeandjustice/datasets/policeforceareadatatables}
}
\usage{
places_low_level_crime
}
\description{
A dataset containing statistics on the mean rate of low-level crime (bicycle
theft and shoplifting) per 1,000, per Welsh Local Authority.
}
\keyword{datasets}
