% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_personal_crime}
\alias{places_personal_crime}
\title{Average rates of Personal Crime per 1k (2023-2024)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{personal_crime_rate_per_1k}{Average rate of low-level-crime (Violence
against the person, sexual offences, and criminal damage and arson)}
\item{year}{September 2023 - September 2024}
...
}

A data frame with 22 rows and 3 variables:
\describe{
\item{ltla25_code}{Local Authority Code}
\item{personal_crime_rate_per_1k}{Average rate of low-level-crime (Violence
against the person, sexual offences, and criminal damage and arson)}
\item{year}{September 2023 - September 2024}
...
}
}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/crimeandjustice/datasets/policeforceareadatatables}

\url{https://www.ons.gov.uk/peoplepopulationandcommunity/crimeandjustice/datasets/policeforceareadatatables}
}
\usage{
places_personal_crime

places_personal_crime
}
\description{
A dataset containing statistics on the mean rate of personal crime (Violence
against the person, sexual offences, and criminal damage and arson) per
1,000, per Welsh Local Authority.

A dataset containing statistics on the mean rate of personal crime (Violence
against the person, sexual offences, and criminal damage and arson) per
1,000, per Welsh Local Authority.
}
\keyword{datasets}
