% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_unemployment}
\alias{places_unemployment}
\title{Percentage of People Experiencing Unemployment (2023-24)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla225_code}{Local Authority Code}
\item{unemployment_percentage}{Percentage of people experiencing unemployment
excluding students. Ages 16-64.}
\item{year}{September 2023 to September 2024}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Business-Economy-and-Labour-Market/People-and-Work/Economic-Inactivity/economicinactivityratesexcludingstudents-by-welshlocalarea-year}
}
\usage{
places_unemployment
}
\description{
A dataset containing statistics on the percentage of people experiencing
unemployment, excluding students.
}
\keyword{datasets}
