% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rank_na_first}
\alias{rank_na_first}
\title{Rank indicators with NAs first (i.e. 1 = worst)}
\usage{
rank_na_first(x)
}
\arguments{
\item{x}{Data to rank}
}
\description{
Rank indicators with NAs first (i.e. 1 = worst)
}
