% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_ranks}
\alias{scale_ranks}
\title{Normalise ranks to a range between 0 and 1}
\usage{
scale_ranks(x)
}
\arguments{
\item{x}{Ranks to normalise}
}
\description{
Normalise ranks to a range between 0 and 1
}
