% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weight_indicators_mfla}
\alias{weight_indicators_mfla}
\title{Weight indicators within a domain using MFLA. This function will calculate
over all numeric variables in a dataframe.}
\usage{
weight_indicators_mfla(data)
}
\arguments{
\item{data}{Data frame containing indicators to be weighted}
}
\description{
Method:
\enumerate{
\item Normalise each indicator to Mean = 0, SD = 1
\item Perform MLFA and extract weights for that domain
\item Multiply model weights by respective column to get weighted indicators
}
}
