% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_cardiovascular_conditions}
\alias{lives_cardiovascular_conditions}
\title{Percentage Rate of Cardiovascular Conditions (CHD, Atrial Fibrillation,
Heart Failure, Stroke & TIA) (2024)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authroity Code}
\item{cardiovascular_conditions_percentage}{Percentage of people who self-reported
as having CHD, Atrial Fibrillation, Heart Failure, and/or Stroke & TIA}
\item{year}{Year}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Health-and-Social-Care/NHS-Primary-and-Community-Activity/GMS-Contract/diseaseregisters-by-localhealthboard}
}
\usage{
lives_cardiovascular_conditions
}
\description{
A dataset containing statistics on the percentage of people who self-reported
as having CHD, Atrial Fibrillation, Heart Failure, and/or Stroke & TIA in
Welsh Council Areas.
}
\details{
To note: StatsWales Disease Prevalence data only captures disease prevalence
in a 'snapshot', or a single point of time. This means the year data
represents when the data was extracted, not the period covered.
}
\keyword{datasets}
