% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lives_high_blood_pressure}
\alias{lives_high_blood_pressure}
\title{Percentage of Adults Diagnosed with High Blood Pressure (2024)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{high_blood_pressure_percentage}{Percentage of adults diagnosed with
high blood pressure}
\item{year}{Year}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Health-and-Social-Care/NHS-Primary-and-Community-Activity/GMS-Contract/diseaseregisters-by-localhealthboard}
}
\usage{
lives_high_blood_pressure
}
\description{
A dataset containing statistics on percentage of adults diagnosed with high
blood pressure, by Council (2024).
}
\details{
To note: StatsWales Disease Prevalence data only captures disease prevalence
in a 'snapshot', or a single point of time. This means the year data
represents when the data was extracted, not the period covered.
}
\keyword{datasets}
