% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_road_safety}
\alias{places_road_safety}
\title{Road safety (2023)}
\format{
A data frame with 22 rows and 3 variables:
\describe{
\item{ltla24_code}{Local Authority Code}
\item{road_accident_count}{Number of people killed or injured, per Local
Authority}
\item{year}{Year}
\item{domain}{Indicator domain - lives, people, or places}
\item{subdomain}{Indicator subdomain}
\item{is_higher_better}{Does higher scores correspond to better outcomes}

...
}
}
\source{
\url{https://statswales.gov.wales/Catalogue/Transport/Roads/Road-Accidents/Casualties/numberofcasualties-by-quarter-year-localauthority-policeforce}
}
\usage{
places_road_safety
}
\description{
A dataset containing number of people killed or injured in each Welsh Council
Area in 2023 (latest revised and checked data available).
}
\keyword{datasets}
